ui분리 및 auth 분리
